Add BMP redundant connection detection logic

+++ bgp/bmp-impl/src/main/java/org/opendaylight/protocol/bmp/impl/app/BmpRouterImpl.java
@@ -77,26 +77,45 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
     public BmpRouterImpl(final RouterSessionManager sessionManager) {
         this.sessionManager = Preconditions.checkNotNull(sessionManager);
         this.domDataBroker = sessionManager.getDomDataBroker();
-        this.domTxChain = sessionManager.getDomDataBroker().createTransactionChain(this);
+        this.domTxChain = this.domDataBroker.createTransactionChain(this);
         this.extensions = sessionManager.getExtensions();
         this.tree = sessionManager.getCodecTree();
     }
 
     @Override
-    public void onSessionUp(final BmpSession session) {
+    public boolean onSessionUp(final BmpSession session) {
         this.session = session;
         this.routerIp = InetAddresses.toAddrString(session.getRemoteAddress());
         this.routerId = new RouterId(Ipv4Util.getIpAddress(session.getRemoteAddress()));
+        // check if this session is redundant
+        if (!this.sessionManager.addSessionListener(this)) {
+            LOG.warn("Redundant session (local <-> {}) detected. This BMP session will be abandoned.", session.getRemoteAddress().getHostAddress());
+            try {
+                this.session.close();
+            } catch (Exception e) {
+                LOG.error("Error occurred when trying to close session.", e);
+            } finally {
+                // ideally the onSessionDown() will be invoked automatically when session.close() is called
+                // session.close() -> channelInactive() -> endOfInput() -> onSessionDown()
+                onSessionDown(this.session, new IllegalStateException("Session is redundant."));
+            }
+            return false;
+        }
         this.routerYangIId = YangInstanceIdentifier.builder(this.sessionManager.getRoutersYangIId()).nodeWithKey(Router.QNAME,
                 ROUTER_ID_QNAME, this.routerIp).build();
         this.peersYangIId = YangInstanceIdentifier.builder(routerYangIId).node(Peer.QNAME).build();
         createRouterEntry();
-        this.sessionManager.addSessionListener(this);
+        return true;
     }
 
     @Override
     public void onSessionDown(final BmpSession session, final Exception e) {
-        LOG.info("Session {} went down.", session);
+        if (isSessionValid()) {
+            // do not display the log when session is invalid
+            LOG.info("Session {} went down.", session);
+            LOG.debug("Exception found", e);
+        }
+        // we still want to tear down as we want to do clean up like closing the transaction chain, etc.
         tearDown();
     }
 
@@ -125,6 +144,8 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
 
     @GuardedBy("this")
     private synchronized void tearDown() {
+        if (this.session == null)   // the session has been teared down before
+            return;
         this.session = null;
         final Iterator<BmpRouterPeer> it = this.peers.values().iterator();
         try {
@@ -136,15 +157,21 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
         } catch(final Exception e) {
             LOG.error("Failed to properly close BMP application.", e);
         } finally {
-            try {
-                final DOMDataWriteTransaction wTx = this.domDataBroker.newWriteOnlyTransaction();
-                wTx.delete(LogicalDatastoreType.OPERATIONAL, this.routerYangIId);
-                wTx.submit().checkedGet();
-            } catch (final TransactionCommitFailedException e) {
-                LOG.error("Failed to remove BMP router data from DS.", e);
+            // remove session only when session is valid, otherwise
+            // we would remove the original valid session when a redundant connection happens
+            // as the routerId is the same for both connection
+            if (isSessionValid()) {
+                try {
+                    // it means the session was closed before it was written to datastore
+                    final DOMDataWriteTransaction wTx = this.domDataBroker.newWriteOnlyTransaction();
+                    wTx.delete(LogicalDatastoreType.OPERATIONAL, this.routerYangIId);
+                    wTx.submit().checkedGet();
+                } catch (final TransactionCommitFailedException e) {
+                    LOG.error("Failed to remove BMP router data from DS.", e);
+                }
+                this.sessionManager.removeSessionListener(this);
             }
         }
-        this.sessionManager.removeSessionListener(this);
     }
 
     @Override
@@ -154,10 +181,15 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
 
     @Override
     public void onTransactionChainSuccessful(final TransactionChain<?, ?> chain) {
-        LOG.debug("Transaction chain {} successfull.", chain);
+        LOG.debug("Transaction chain {} successfully.", chain);
+    }
+
+    private boolean isSessionValid() {
+        return (this.routerYangIId != null);
     }
 
     private void createRouterEntry() {
+        Preconditions.checkArgument(isSessionValid());
         final DOMDataWriteTransaction wTx = this.domTxChain.newWriteOnlyTransaction();
         wTx.put(LogicalDatastoreType.OPERATIONAL, this.routerYangIId,
                 Builders.mapEntryBuilder()
@@ -169,6 +201,7 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
     }
 
     private void onInitiate(final InitiationMessage initiation) {
+        Preconditions.checkArgument(isSessionValid());
         final DOMDataWriteTransaction wTx = this.domTxChain.newWriteOnlyTransaction();
         wTx.merge(LogicalDatastoreType.OPERATIONAL, this.routerYangIId,
                 Builders.mapEntryBuilder()
